Allow more complex content types.
We use Content-type headers that look like this:

Content-Type: application/vnd.api.v1+json; charset=utf-8

and curently slumber barfs on these, refusing to decode.  This change
allows you to specify, say, format='json' when initializing the API
object and deserialize properly.

I'm not sure if this is the totally right spot for this code, but I wanted
to bring this to ya'lls attention.